+++ packages/react-native/src/evm/providers/thirdweb-provider.tsx
@@ -52,7 +52,7 @@ interface ThirdwebProviderProps<TChains extends Chain[]>
    * />
    * ```
    *
-   * * ```jsx
+   * ```jsx
    * import { en } from "@thirdweb-dev/react-native";
    *
    * <ThirdwebProvider
@@ -86,6 +86,7 @@ interface ThirdwebProviderProps<TChains extends Chain[]>
  * @example
  * You can wrap your application with the provider as follows:
  *
+ * ```jsx
  * import { ThirdwebProvider } from "@thirdweb-dev/react-native";
  *
  * const App = () => {
@@ -95,7 +96,7 @@ interface ThirdwebProviderProps<TChains extends Chain[]>
  *     </ThirdwebProvider>
  *   );
  * };
- *
+ * ```
  */
 export const ThirdwebProvider = <
   TChains extends Chain[] = typeof defaultChains,