+++ packages/wallets/src/evm/connectors/embedded-wallet/index.ts
@@ -89,7 +89,7 @@ export class EmbeddedWalletConnector extends Connector<EmbeddedWalletConnectionA
     if (!this.user) {
       throw new Error("Embedded Wallet is not connected");
     }
-    return this.user.walletAddress;
+    return await this.getSigner().then((signer) => signer.getAddress());
   }
 
   async isConnected(): Promise<boolean> {
@@ -115,7 +115,7 @@ export class EmbeddedWalletConnector extends Connector<EmbeddedWalletConnectionA
     }
 
     const user = await this.getUser();
-    const signer = await user?.wallet.getEthersJsSigner({
+    const signer = await user.wallet.getEthersJsSigner({
       rpcEndpoint: this.options.chain.rpc[0] || "", // TODO: handle chain.rpc being empty array
     });
 
@@ -138,15 +138,17 @@ export class EmbeddedWalletConnector extends Connector<EmbeddedWalletConnectionA
       throw new Error("Chain not configured");
     }
 
-    // update chain in wallet
-    await this.user?.wallet.setChain({ chain: "Ethereum" }); // just pass Ethereum no matter what chain we are going to connect
-
-    // update signer
-    this.#signer = await this.user?.wallet.getEthersJsSigner({
-      rpcEndpoint: chain.rpc[0] || "", // TODO: handle chain.rpc being empty array
-    });
-
-    this.emit("change", { chain: { id: chainId, unsupported: false } });
+    try {
+      // update chain in wallet
+      await this.user?.wallet.setChain({ chain: "Ethereum" }); // just pass Ethereum no matter what chain we are going to connect
+      // update signer
+      this.#signer = await this.user?.wallet.getEthersJsSigner({
+        rpcEndpoint: chain.rpc[0] || "",
+      });
+      this.emit("change", { chain: { id: chainId, unsupported: false } });
+    } catch (e) {
+      console.warn("Failed to switch chain", e);
+    }
   }
 
   async setupListeners() {
@@ -178,11 +180,11 @@ export class EmbeddedWalletConnector extends Connector<EmbeddedWalletConnectionA
     this.emit("disconnect");
   };
 
-  async getUser(): Promise<InitializedUser | null> {
+  private async getUser(): Promise<InitializedUser> {
     if (
       !this.user ||
       !this.user.wallet ||
-      !this.user.wallet.getEthersJsSigner
+      !this.user.wallet.getEthersJsSigner // when serializing, functions are lost, need to rehydrate
     ) {
       const embeddedWalletSdk = this.getEmbeddedWalletSDK();
       const user = await embeddedWalletSdk.getUser();
@@ -191,27 +193,25 @@ export class EmbeddedWalletConnector extends Connector<EmbeddedWalletConnectionA
           this.user = user;
           break;
         }
+        default: {
+          // if logged out or unitialized, we can't get a signer, so throw an error
+          throw new Error(
+            "Embedded Wallet is not authenticated, please authenticate first",
+          );
+        }
       }
     }
     return this.user;
   }
 
   async getEmail() {
-    // implicit call to set the user
-    await this.getSigner();
-    if (!this.user) {
-      throw new Error("No user found, Embedded Wallet is not connected");
-    }
-    return this.user.authDetails.email;
+    const user = await this.getUser();
+    return user.authDetails.email;
   }
 
   async getRecoveryInformation() {
-    // implicit call to set the user
-    await this.getSigner();
-    if (!this.user) {
-      throw new Error("No user found, Embedded Wallet is not connected");
-    }
-    return this.user.authDetails;
+    const user = await this.getUser();
+    return user.authDetails;
   }
 
   async sendVerificationEmail({
@@ -235,7 +235,7 @@ export class EmbeddedWalletConnector extends Connector<EmbeddedWalletConnectionA
         });
       }
       case "apple":
-      // case "facebook":
+      case "facebook":
       case "google": {
         const oauthProvider = oauthStrategyToAuthProvider[strategy];
         return ewSDK.auth.loginWithOauth({
@@ -274,6 +274,6 @@ const oauthStrategyToAuthProvider: Record<
   AuthProvider
 > = {
   google: AuthProvider.GOOGLE,
-  // facebook: AuthProvider.FACEBOOK,
+  facebook: AuthProvider.FACEBOOK,
   apple: AuthProvider.APPLE,
 };