+++ packages/wallets/src/evm/connectors/magic/types.ts
@@ -22,17 +22,17 @@ export interface MagicOptions {
 export interface MagicAuthOptions extends MagicOptions {
   /**
    * enable email login for the Magic Auth.
-   * @default true
+   * @defaultValue true
    */
   emailLogin?: boolean;
   /**
    * enable sms login for the Magic Auth.
-   * @default true
+   * @defaultValue true
    */
   smsLogin?: boolean;
   /**
    * Specify whether to use the Magic Auth or Magic Connect.
-   * @default "auth"
+   * @defaultValue "auth"
    */
   type?: "auth" | "connect";
 
@@ -52,7 +52,7 @@ export interface MagicConnectorBaseOptions {
   magicSdkConfiguration?: MagicSDKAdditionalConfiguration;
   /**
    * Specify whether to use the Magic Auth or Magic Connect.
-   * @default "auth"
+   * @defaultValue "auth"
    */
   type?: "auth" | "connect";
 }