+++ packages/wallets/src/evm/wallets/base.ts
@@ -41,6 +41,7 @@ export abstract class AbstractClientWallet<
   protected options?: WalletOptions<TAdditionalOpts>;
   static meta: WalletMeta;
   #connectParams: ConnectParams<TConnectParams> | undefined;
+
   getMeta() {
     return (this.constructor as typeof AbstractClientWallet).meta;
   }