+++ test/runtimes/node-module/fixtures/sdk.js
@@ -1,0 +1,3 @@
+import { ThirdwebSDK } from "@thirdweb-dev/sdk";
+
+new ThirdwebSDK(1);